Replace `.sr-only` class with `.visually-hidden`
Bootstrap 5 has switched the "screen reader" class from `.sr-only` to
`.visually-hidden`. Update the library so that form labels which are
set to be hidden are actually hidden while stile being available for
assistive technologies.

This change is not currently backwards compatible, but could easily be
made so by adding and extra check to renderers line #474 for the old
'sr-only' value.