Remove `.form-group` and use `.mb-3`
Bootstrap 5 removes the `.form-group` class in preference of using the
utility class `.mb-3`, or another of the utility classes which fits the
use case better. Replace the former with the latter in the code.

This change in Bootstrap seems to be aiming to make form groups more
flexible - this code should not have a problem setting `.mb-3` since it
is only a default and can be overridden by using the `form_group_class`
parameter.