Update select and file widget classes
The file upload field in older versions of Bootstrap needed its own
special class. Bootstrap 5 changed that and allowed it to work with the
default `.form-control` class. Remove its special treatment.

The select widget used to work with the default `.form-control` class,
but was updated to use `.form-select`. Make an exception for choice
fields.